v1.0.1

+++ CLEOPlus/Misc.cpp
@@ -1529,7 +1529,7 @@ OpcodeResult WINAPI DONT_DELETE_CHAR_UNTIL_TIME(CScriptThread* thread)
 {
 	CPed *ped = CPools::GetPed(CLEO_GetIntOpcodeParam(thread));
 	unsigned int time = CLEO_GetIntOpcodeParam(thread) + CTimer::m_snTimeInMilliseconds;
-	if (ped->field_54C < time) { // m_nTimeTillWeNeedThisPed
+	if ((unsigned int)ped->field_54C < time) { // m_nTimeTillWeNeedThisPed
 		ped->field_54C = time;
 	}
 	return OR_CONTINUE;
